backend/requirements.txt
# Keep dependencies minimal and CPU-only.
# PyTorch is required only because sentence-transformers depends on it.
# This forces the CPU wheel instead of pulling NVIDIA/CUDA packages.
--extra-index-url https://download.pytorch.org/whl/cpu

# Core backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
structlog==24.4.0
email-validator==2.2.0

# Database (needed for upcoming Week 10 work)
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.1.18

# LLM provider actually used by the project
groq==0.4.1

# Semantic search
numpy==1.26.2
regex==2023.10.3
torch==2.2.2+cpu
sentence-transformers==2.3.1
faiss-cpu>=1.8.0

# Scripts / demos
requests>=2.31.0,<3.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
pytest-asyncio==0.23.8